+++ changelog.md
@@ -1,3 +1,7 @@
+## v1.0.8
+- change params verifyDomain to verifyMx
+- change params verifyMailbox to verifySmtp
+
 ## v1.0.7
 - Add PSL to support ( isValidEmailDomain )
 - Refactor tests